+++ text-based-adventure-i.py
@@ -14,12 +14,17 @@ with open("easy_campaign_gear.csv", newline="") as easy_gear:
     for row in read_easy_gear:
         aux_gear = sc.Gear(str(row[0]), int(row[1]), int(row[2]))
         gear_to_unlock.append(aux_gear)
-for element in gear_to_unlock[1:]:
-    print(element)
 
-first_player = sc.Warrior(100, 10, 10, "Player 1")
-first_monster = sc.Warrior(10, 150, 9, "First monster created")
-second_monster = sc.Warrior(20, 15, 9, "Second monster created")
+#Create monsters to battle
+monsters_to_battle = [sc.Warrior]
+with open("easy_campaign_monsters.csv", newline="") as easy_monsters:
+    read_easy_monsters = csv.reader(easy_monsters)
+    next(read_easy_monsters)
+    for row in read_easy_monsters:
+        aux_monster = sc.Warrior(str(row[0]), int(row[1]), int(row[2]), int(row[3]))
+        monsters_to_battle.append(aux_monster)
+
+first_player = sc.Warrior("Player 1", 100, 10, 10)
 first_item = sc.Item("hp_item", 100, "primer item")
 second_item = sc.Item("atk_item", 100, "primer item")
 third_item = sc.Item("def_item", 100, "primer item")
@@ -30,12 +35,10 @@ gear_item_1 = sc.Gear("Spear of death", 5, -5,)
 gear_item_2 = sc.Gear("Shield of power", -5, 5)
 first_player.add_to_gear_bag(gear_item_1)
 first_player.add_to_gear_bag(gear_item_2)
-first_player.print_gear_bag()
-#first_player.bag_items_print()
-#first_player.equip_item(first_item)
-#first_player.select_items()
-#ts.turn(first_player, first_monster)
-#print(first_player.health_stat)
-#ts.turn(first_player, second_monster)
-#print(first_player.health_stat)
+
+monsters_index = 1
+while first_player.is_alive():
+    ts.turn(first_player, monsters_to_battle[monsters_index])
+    monsters_index += 1
+